renames header files.
