+++ internal/phpheaders/phpheaders_test.go
@@ -1,4 +1,4 @@
-package headers
+package phpheaders
 
 import (
 	"testing"
